pyproject.toml
[project]
name = "ad-generator"
version = "0.1.0"
description = "Generate video advertisements from product pages using Claude Agent SDK and FreePik Veo 3"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "anyio>=4.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
]

[project.scripts]
ad-generator = "ad_generator.main:cli"
ad-generator-server = "ad_generator.api:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ad_generator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
